requirements.txt
discord.py==2.5.2
yt-dlp==2025.7.21
requests==2.32.4
python-dotenv==1.1.1
aiosqlite==0.20.0
aiofiles==24.1.0
urllib3==2.2.2
aiohttp==3.9.1
